requirements.txt
# Alpaca SDK for trading
alpaca-py>=0.21.0

# Technical analysis library
pandas-ta>=0.3.14b

# Backtesting library
backtrader>=1.9.78.123

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Configuration management
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
